# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "接口文档",
        "title": "go-server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/private/v1/:id/attributes": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类属性"
                ],
                "summary": "分类参数更新",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/attributedto.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类属性"
                ],
                "summary": "分类参数创建",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/attributedto.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/attributedto.PublicDTO"
                        }
                    }
                }
            }
        },
        "/api/private/v1/categories": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品分类"
                ],
                "summary": "商品分类列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "1",
                            "2",
                            "3"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "CategoryTypeLevel1",
                            "CategoryTypeLevel2",
                            "CategoryTypeLevel3"
                        ],
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/category.PublicDTO"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品分类"
                ],
                "summary": "商品分类更新",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/category.PrivateDTO"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品分类"
                ],
                "summary": "商品分类创建",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/category.PrivateDTO"
                        }
                    }
                }
            }
        },
        "/api/private/v1/categories/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品分类"
                ],
                "summary": "商品分类删除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/categories/:id/attributes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类属性"
                ],
                "summary": "分类参数列表",
                "parameters": [
                    {
                        "enum": [
                            "only",
                            "many"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "AttrSelOnly",
                            "AttrSelMany"
                        ],
                        "description": "属性录入方式",
                        "name": "sel",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/attributedto.PublicDTO"
                            }
                        }
                    }
                }
            }
        },
        "/categories/:id/attributes/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类属性"
                ],
                "summary": "分类参数删除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "attributedto.CreateRequest": {
            "type": "object",
            "required": [
                "attr_name",
                "attr_sel"
            ],
            "properties": {
                "attr_name": {
                    "description": "属性名称",
                    "type": "string"
                },
                "attr_sel": {
                    "description": "属性录入方式",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AttrSel"
                        }
                    ]
                }
            }
        },
        "attributedto.PublicDTO": {
            "type": "object",
            "properties": {
                "attr_id": {
                    "description": "属性id",
                    "type": "integer"
                },
                "attr_name": {
                    "description": "属性名称",
                    "type": "string"
                },
                "attr_sel": {
                    "description": "属性录入方式",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AttrSel"
                        }
                    ]
                },
                "attr_vals": {
                    "description": "可选值",
                    "type": "string"
                },
                "attr_write": {
                    "description": "属性录入方式",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AttrWrite"
                        }
                    ]
                },
                "cat_id": {
                    "description": "分类id",
                    "type": "integer"
                }
            }
        },
        "attributedto.UpdateRequest": {
            "type": "object",
            "required": [
                "attr_name",
                "attr_sel"
            ],
            "properties": {
                "attr_name": {
                    "description": "属性名称",
                    "type": "string"
                },
                "attr_sel": {
                    "description": "属性录入方式",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AttrSel"
                        }
                    ]
                },
                "attr_vals": {
                    "type": "string"
                }
            }
        },
...
        "category.CreateRequest": {
            "type": "object",
            "required": [
                "cat_name"
            ],
            "properties": {
                "cat_level": {
                    "description": "分类层级",
                    "type": "integer"
                },
                "cat_name": {
                    "description": "分类名称",
                    "type": "string"
                },
                "cat_pid": {
                    "description": "父ID",
                    "type": "integer"
                }
            }
        },
        "category.PrivateDTO": {
            "type": "object"
        },
        "category.PublicDTO": {
            "type": "object",
            "properties": {
                "cat_deleted": {
                    "description": "是否删除",
                    "type": "boolean"
                },
                "cat_id": {
                    "description": "分类ID",
                    "type": "integer"
                },
                "cat_level": {
                    "description": "分类层级",
                    "type": "integer"
                },
                "cat_name": {
                    "description": "分类名称",
                    "type": "string"
                },
                "cat_pid": {
                    "description": "父ID",
                    "type": "integer"
                },
                "children": {
                    "description": "omitempty：如果字段是“空值”，就不返回",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/category.PublicDTO"
                    }
                }
            }
        },
        "category.UpdateRequest": {
            "type": "object",
            "properties": {
                "cat_name": {
                    "description": "分类名称",
                    "type": "string"
                }
            }
        },
...
        "model.AttrSel": {
            "type": "string",
            "enum": [
                "only",
                "many"
            ],
            "x-enum-varnames": [
                "AttrSelOnly",
                "AttrSelMany"
            ]
        },
        "model.AttrWrite": {
            "type": "string",
            "enum": [
                "manual",
                "list"
            ],
            "x-enum-varnames": [
                "AttrWriteManual",
                "AttrWriteList"
            ]
        },
...
    }
}